+++ src/elements/Variant.php
@@ -745,7 +745,7 @@ class Variant extends Purchasable implements NestedElementInterface
         if ($url = parent::getUrl()) {
             return $url;
         }
-        
+
         // Default URL is the product's URL with the variant ID as a query parameter
         $productUrl = $this->getOwner()?->getUrl();
         return $productUrl ? UrlHelper::urlWithParams($productUrl, ['variant' => $this->id]) : null;